discard worker and create new one if heic conversion fails
