+++ src/utils/comlink/index.ts
@@ -6,17 +6,6 @@ export interface ComlinkWorker {
     worker: Worker;
 }
 
-export const getDedicatedConvertWorker = (): ComlinkWorker => {
-    if (runningInBrowser()) {
-        const worker = new Worker(
-            new URL('worker/convert.worker.js', import.meta.url),
-            { name: 'ente-convert-worker' }
-        );
-        const comlink = Comlink.wrap(worker);
-        return { comlink, worker };
-    }
-};
-
 const getDedicatedFFmpegWorker = (): ComlinkWorker => {
     if (runningInBrowser()) {
         const worker = new Worker(
